requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langgraph

chromadb
sentence-transformers

numpy
scikit-learn

requests
tqdm

pytest
rich